+++ lib/src/database/table_entry.dart
@@ -1,26 +1,27 @@
-
-enum SelectType {
-  and,
-  or
-}
-class SelectEntry {
-  final String columnName;
-  final String value;
-  final SelectType type;
-
-  SelectEntry(this.columnName, this.value, this.type);
-
-  SelectEntry.and(String columnName, String value): this(columnName, value, SelectType.and);
-  SelectEntry.or(String columnName, String value): this(columnName, value, SelectType.or);
-}
-
 abstract class TableData<T> {
   late String tableName;
+
   T fromJson(Map<String, dynamic> json);
 }
 
-abstract class TableEntry<T> {
+mixin TableEntry<T> {
   abstract int? id;
+
   Map<String, dynamic> toJson();
-  List<Map<String, dynamic>> addUserId(String userId);
+
+  T addUserId(String userId);
 }
+
+mixin HasId {
+  int? get tableId;
+}
+
+mixin HasUserId {
+  String? get userId;
+}
+
+mixin hasCopyWith<T> {
+  T get copyWith;
+}
+
+typedef TableEntryCreator<T> = TableEntry<T> Function(T data);