+++ lib/supa_manager.dart
@@ -8,3 +8,4 @@ export 'src/auth/login_state.dart';
 export 'src/auth/result.dart';
 export 'src/database/supa_database_manager.dart';
 export 'src/database/table_entry.dart';
+export 'src/database/select_entry.dart';